+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 
 ### Enhancements
 
+* Add `device.freeMemory` to OOM and Thermal Kill events on iOS 13 and later.
+  [#1357](https://github.com/bugsnag/bugsnag-cocoa/pull/1357)
+
 * Add `device.time` to OOM and Thermal Kill events.
   [#1355](https://github.com/bugsnag/bugsnag-cocoa/pull/1355)